+++ src/main/java/ph/devcon/rapidpass/services/ApproverAuthService.java
@@ -3,26 +3,31 @@ package ph.devcon.rapidpass.services;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.codec.DecoderException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 import ph.devcon.rapidpass.config.JwtSecretsConfig;
 import ph.devcon.rapidpass.entities.Registrar;
 import ph.devcon.rapidpass.entities.RegistrarUser;
+import ph.devcon.rapidpass.enums.RegistrarUserStatus;
+import ph.devcon.rapidpass.kafka.RegistrarUserRequestProducer;
 import ph.devcon.rapidpass.models.AgencyAuth;
 import ph.devcon.rapidpass.models.AgencyUser;
 import ph.devcon.rapidpass.repositories.RegistrarRepository;
 import ph.devcon.rapidpass.repositories.RegistrarUserRepository;
 import ph.devcon.rapidpass.utilities.CryptUtils;
 import ph.devcon.rapidpass.utilities.JwtGenerator;
+import ph.devcon.rapidpass.validators.StandardDataBindingValidation;
+import ph.devcon.rapidpass.validators.entities.agencyuser.BaseAgencyUserRequestValidator;
+import ph.devcon.rapidpass.validators.entities.agencyuser.NewSingleAgencyUserRequestValidator;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
+import java.util.UUID;
 
 import static ph.devcon.rapidpass.utilities.CryptUtils.passwordCompare;
 import static ph.devcon.rapidpass.utilities.CryptUtils.passwordHash;
@@ -39,14 +44,20 @@ public class ApproverAuthService {
     private final RegistrarUserRepository registrarUserRepository;
     private final RegistrarRepository registrarRepository;
     private final JwtSecretsConfig jwtSecretsConfig;
+    private final RegistrarUserRequestProducer registrarUserRequestProducer;
+
+    @Value("${kafka.enabled:false}")
+    protected boolean isKafaEnabled;
 
     @Autowired
     public ApproverAuthService(final RegistrarUserRepository registrarUserRepository,
                                final RegistrarRepository registrarRepository,
-                               final JwtSecretsConfig jwtSecretsConfig) {
+                               final JwtSecretsConfig jwtSecretsConfig,
+                               final RegistrarUserRequestProducer registrarUserRequestProducer) {
         this.registrarUserRepository = registrarUserRepository;
         this.registrarRepository = registrarRepository;
         this.jwtSecretsConfig = jwtSecretsConfig;
+        this.registrarUserRequestProducer = registrarUserRequestProducer;
     }
 
     /**
@@ -88,28 +99,43 @@ public class ApproverAuthService {
      * @throws InvalidKeySpecException this is returned when the hashing algorithm fails. This is an illegal state
      * @throws NoSuchAlgorithmException this is returned when the hashing algorithm fails. This is an illegal state
      */
-    public final void createAgencyCredentials(final AgencyUser user) throws IllegalArgumentException, InvalidKeySpecException, NoSuchAlgorithmException {
-        final String registrarShortName = user.getRegistrar();
-        final Registrar registrar = registrarRepository.findByShortName(registrarShortName);
-        if (registrar == null) {
-            log.error("the registrar provided does not exist");
-            throw new IllegalArgumentException("unable to find registrar for user");
+    public final RegistrarUser createAgencyCredentials(final AgencyUser user) throws IllegalArgumentException, InvalidKeySpecException, NoSuchAlgorithmException {
+
+        if (!user.isBatchUpload()) {
+            BaseAgencyUserRequestValidator validator = new NewSingleAgencyUserRequestValidator(this.registrarUserRepository, this.registrarRepository);
+            StandardDataBindingValidation dataValidator = new StandardDataBindingValidation(validator);
+            dataValidator.validate(user);
         }
 
-        final RegistrarUser registrarUser = this.registrarUserRepository.findByUsername(user.getUsername());
-        if (registrarUser != null) {
-            log.error("user already exists");
-            throw new IllegalArgumentException("user already exists");
+        Registrar registrar = this.registrarRepository.findByShortName(user.getRegistrar());
+
+        final RegistrarUser registrarUser = new RegistrarUser();
+
+        registrarUser.setRegistrarId(registrar);
+        registrarUser.setUsername(user.getUsername());
+
+
+        if (user.isBatchUpload()) {
+            registrarUser.setFirstName(user.getFirstName());
+            registrarUser.setLastName(user.getLastName());
+            registrar.setEmail(user.getEmail());
+            registrarUser.setStatus(RegistrarUserStatus.INACTIVE.toString());
+
+            // Uses a v4 UUID.
+            String randomUniqueActivationCode = UUID.randomUUID().toString();
+            registrarUser.setAccessKey(randomUniqueActivationCode);
+
+            if (isKafaEnabled)
+                registrarUserRequestProducer.sendMessage(registrarUser.getUsername(), registrarUser);
+
+        } else if (user.isIndividualRegistration()) {
+            final String hashedPassword = passwordHash(user.getPassword());
+            registrarUser.setPassword(hashedPassword);
+            registrarUser.setStatus(RegistrarUserStatus.ACTIVE.toString());
         }
 
-        final RegistrarUser newRegistrarUser = new RegistrarUser();
-        newRegistrarUser.setRegistrarId(registrar);
-        newRegistrarUser.setUsername(user.getUsername());
-        final String hashedPassword = passwordHash(user.getPassword());
-        newRegistrarUser.setPassword(hashedPassword);
-        newRegistrarUser.setStatus("active");
-
-        registrarUserRepository.save(newRegistrarUser);
+        registrarUserRepository.saveAndFlush(registrarUser);
+        return registrarUser;
     }
 
     /**